+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Rename `Code` to `CodeBlock`
+- Support CSS color in inline code
 
 ## 0.20.1